feat: add randome speed work for cash
